requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
bcrypt==3.1.6
blinker==1.4
boto3==1.9.208
botocore==1.12.223
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dnspython==1.16.0
docutils==0.15.2
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.18.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
mongoengine==0.18.0
passlib==1.7.1
phonenumbers==8.10.14
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.10
stripe==2.35.1
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1

